automatic-anamoly-detection/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# PyTorch CPU wheels known to install cleanly on Spaces
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

anomalib==2.1.0
timm==1.0.8
omegaconf==2.3.0
opencv-python-headless==4.10.0.84
pillow==10.4.0
numpy==1.26.4
scikit-learn==1.4.2
scikit-image==0.24.0
open_clip_torch==2.24.0
requests>=2.31.0
gradio==4.44.1
